Mov auth0 and apollo to their corresponding services
