+++ src/services/auth/AuthWrapper.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { useEffect, useState } from 'react';
 import { useAuth0 } from '@auth0/auth0-react';
 import { Center, Spinner } from '@chakra-ui/react';
